semiterminado

+++ src/main/java/com/example/demo/service/MatriculaService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface MatriculaService {
     List<Matricula> getAllMatriculas();
     void addMatricula(Matricula matricula);
-    Matricula getMatriculaById(int id);
-    void deleteMatricula(int id);
+    Matricula getMatriculaById(Long id);
+    void deleteMatricula(Long id);
     void updateMatricula(Matricula matricula);
 }
